+++ doers/src/zone/cloudinit.rs
@@ -1,8 +1,8 @@
-use crate::zone::config::GurpZoneBhyve;
+use crate::zone::config::{CloudInitConfig, GurpZoneFilesystem};
 use anyhow::{Context, ensure};
 use camino::{Utf8Path, Utf8PathBuf};
 use camino_tempfile::Utf8TempDir;
-use common::constants::MKISOFS_BIN;
+use common::constants::{MKISOFS_BIN, ZONECFG_BIN};
 use common::info::dump_config;
 use common::types::ApplyOpts;
 use serde_json::Value;
@@ -19,11 +19,11 @@ use uuid::Uuid;
 // The user can give us files to copy in, and also structs that we can turn into YAML on their
 // behalf.
 
-pub fn iso_path(uuid: &Uuid) -> Utf8PathBuf {
-    Utf8PathBuf::from("/tmp").join(format!("{uuid}.iso"))
-}
-
-pub fn setup(config: &GurpZoneBhyve, iso_file: &Utf8Path, opts: &ApplyOpts) -> anyhow::Result<()> {
+pub fn setup(
+    config: &CloudInitConfig,
+    iso_file: &Utf8Path,
+    opts: &ApplyOpts,
+) -> anyhow::Result<()> {
     tracing::debug!("Setting up Cloudinit");
 
     ensure!(
@@ -40,53 +40,72 @@ pub fn setup(config: &GurpZoneBhyve, iso_file: &Utf8Path, opts: &ApplyOpts) -> a
     Ok(())
 }
 
+pub fn remove(zone: &str) -> anyhow::Result<()> {
+    tracing::debug!("removing cloudinit cdrom from zone config");
+    // It's safe to do this here. The config won't be re-read until the zone
+    // boots
+    let _ = cmd_output!(ZONECFG_BIN, "-z", zone, "remove attr name=cdrom")
+        .with_context(|| format!("failed to remove cdrom attr from zone {}", zone))?;
+
+    tracing::debug!("removing cloudinit lofs from zone config");
+    let _ = cmd_output!(ZONECFG_BIN, "-z", zone, "remove fs type=lofs")
+        .with_context(|| format!("failed to remove cdrom lofs from zone {}", zone))?;
+
+    Ok(())
+}
+
+pub fn iso_path(uuid: &Uuid) -> Utf8PathBuf {
+    Utf8PathBuf::from("/tmp").join(format!("{uuid}.iso"))
+}
+
+pub fn zone_config_snippet(uuid: &Uuid) -> String {
+    let iso_path = iso_path(uuid);
+
+    let fs = zone_fs!(GurpZoneFilesystem {
+        dir: iso_path.clone(),
+        special: iso_path.clone(),
+        fs_type: "lofs".to_owned(),
+        options: Some(vec!["ro".to_owned()])
+    });
+
+    format!("{}\n{fs}", zone_attr!("cdrom", "string", iso_path))
+}
+
 fn populate(
     build_dir: &Utf8TempDir,
-    config: &GurpZoneBhyve,
+    config: &CloudInitConfig,
     opts: &ApplyOpts,
 ) -> anyhow::Result<()> {
     tracing::debug!("Constructing cloud-init CDROM");
 
-    if let Some(cloudinit_files) = &config.cloudinit_files {
-        copy_files(cloudinit_files, build_dir)?;
+    for (file_name, source) in &config.from {
+        copy_file(source, &build_dir.path().join(file_name))?;
     }
 
-    if let Some(cloudinit_struct) = &config.cloudinit_struct
-        && let Some(obj) = cloudinit_struct.as_object()
-    {
-        for (file_name, content) in obj {
-            struct_to_file(file_name, build_dir, content, opts)?;
-        }
+    for (file_name, content) in &config.from_struct {
+        struct_to_file(file_name, build_dir.path(), content, opts)?;
     }
 
     Ok(())
 }
 
-fn copy_files(cloudinit_files: &Vec<Utf8PathBuf>, build_dir: &Utf8TempDir) -> anyhow::Result<()> {
-    for src_path in cloudinit_files {
-        let basename = src_path
-            .file_name()
-            .with_context(|| format!("Cannot get basename of {}", src_path))?;
+fn copy_file(src: &Utf8Path, dest: &Utf8Path) -> anyhow::Result<()> {
+    tracing::debug!("Copying cloudinit {} to {}", src, dest);
 
-        let target_path = build_dir.path().join(basename);
-
-        tracing::debug!("Copying cloudinit {} to {}", src_path, target_path);
-
-        fs::copy(src_path, &target_path)
-            .with_context(|| format!("failed to copy from {src_path} to {target_path}"))?;
-    }
+    fs::copy(src, dest).with_context(|| format!("failed to copy from {src} to {dest}"))?;
 
     Ok(())
 }
 
 fn struct_to_file(
     file_name: &str,
-    build_dir: &Utf8TempDir,
+    build_dir_path: &Utf8Path,
     content_struct: &Value,
     opts: &ApplyOpts,
 ) -> anyhow::Result<()> {
-    let file_path = build_dir.path().join(file_name);
-    tracing::debug!("Creating cloudinit file {}", file_path);
+    let target = build_dir_path.join(file_name);
+    tracing::debug!("Creating cloudinit file {}", target);
+
     let yaml = serde_yaml_bw::to_string(&content_struct)?;
 
     let cloudinit_content = if file_name == "user-data" {
@@ -106,8 +125,8 @@ fn struct_to_file(
         );
     }
 
-    fs::write(&file_path, cloudinit_content)
-        .with_context(|| format!("failed to write Cloudinit content to {file_path}"))?;
+    fs::write(&target, cloudinit_content)
+        .with_context(|| format!("failed to write Cloudinit content to {target}"))?;
 
     Ok(())
 }